social_media/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.5.0
psycopg2==2.9.9